templated cluster tools for computing non-zero-suppressed shower shaped from PFlow clusters
