+++ RecoEcal/EgammaClusterProducers/src/PFECALSuperClusterProducer.cc
@@ -22,7 +22,6 @@
 
 #include "DataFormats/VertexReco/interface/Vertex.h"
 #include "DataFormats/EcalRecHit/interface/EcalRecHit.h"
-#include "RecoEcal/EgammaCoreTools/interface/EcalClusterTools.h"
 
 #include "DataFormats/Common/interface/ValueMap.h"
